# docs/implementation/evidence/ENG-014/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HoldoutManifestCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "access_scope": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Access Scope",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "object_digest": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Object Digest",
            "type": "string"
          },
          "object_length": {
            "minimum": 0.0,
            "title": "Object Length",
            "type": "integer"
          },
          "overlap_review_digest": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlap Review Digest"
          },
          "protocol_version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Protocol Version",
            "type": "string"
          },
          "retention_class": {
            "default": "official",
            "enum": [
              "official",
              "development",
              "ephemeral"
            ],
            "title": "Retention Class",
            "type": "string"
          },
          "split": {
            "default": "official",
            "enum": [
              "official",
              "development"
            ],
            "title": "Split",
            "type": "string"
          },
          "storage_uri": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Storage Uri",
            "type": "string"
          },
          "task_revision_id": {
            "format": "uuid",
            "title": "Task Revision Id",
            "type": "string"
          }
        },
        "required": [
          "task_revision_id",
          "storage_uri",
          "object_digest",
          "object_length",
          "protocol_version",
          "access_scope"
        ],
        "title": "HoldoutManifestCreateRequest",
        "type": "object"
      },
      "HoldoutManifestSummary": {
        "additionalProperties": false,
        "properties": {
          "access_scope": {
            "title": "Access Scope",
            "type": "string"
          },
          "approved_review_kinds": {
            "items": {
              "type": "string"
            },
            "title": "Approved Review Kinds",
            "type": "array"
          },
          "created_by_user_id": {
            "format": "uuid",
            "title": "Created By User Id",
            "type": "string"
          },
          "evaluator_revision_digest": {
            "title": "Evaluator Revision Digest",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "family_id": {
            "title": "Family Id",
            "type": "string"
          },
          "frozen_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frozen At"
          },
          "frozen_by_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frozen By User Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "manifest_digest": {
            "title": "Manifest Digest",
            "type": "string"
          },
          "object_digest": {
            "title": "Object Digest",
            "type": "string"
          },
          "object_length": {
            "title": "Object Length",
            "type": "integer"
          },
          "overlap_review_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlap Review Digest"
          },
          "protocol_version": {
            "title": "Protocol Version",
            "type": "string"
          },
          "retention_class": {
            "enum": [
              "official",
              "development",
              "ephemeral"
            ],
            "title": "Retention Class",
            "type": "string"
          },
          "split": {
            "enum": [
              "official",
              "development"
            ],
            "title": "Split",
            "type": "string"
          },
          "status": {
            "enum": [
              "draft",
              "reviewed",
              "frozen",
              "retired"
            ],
            "title": "Status",
            "type": "string"
          },
          "storage_uri": {
            "title": "Storage Uri",
            "type": "string"
          },
          "task_revision_id": {
            "format": "uuid",
            "title": "Task Revision Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "task_revision_id",
          "family_id",
          "evaluator_revision_digest",
          "storage_uri",
          "object_digest",
          "object_length",
          "manifest_digest",
          "protocol_version",
          "access_scope",
          "split",
          "retention_class",
          "status",
          "created_by_user_id"
        ],
        "title": "HoldoutManifestSummary",
        "type": "object"
      },
      "HoldoutReviewRequest": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "approve",
              "reject",
              "inconclusive"
            ],
            "title": "Decision",
            "type": "string"
          },
          "evidence_digest": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Evidence Digest",
            "type": "string"
          },
          "reason": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "report": {
            "additionalProperties": true,
            "title": "Report",
            "type": "object"
          },
          "review_kind": {
            "enum": [
              "overlap",
              "storage",
              "isolation",
              "manifest"
            ],
            "title": "Review Kind",
            "type": "string"
          }
        },
        "required": [
          "review_kind",
          "decision",
          "evidence_digest",
          "reason"
        ],
        "title": "HoldoutReviewRequest",
        "type": "object"
      },
      "HoldoutReviewSummary": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "enum": [
              "approve",
              "reject",
              "inconclusive"
            ],
            "title": "Decision",
            "type": "string"
          },
          "evidence_digest": {
            "title": "Evidence Digest",
            "type": "string"
          },
          "holdout_manifest_id": {
            "format": "uuid",
            "title": "Holdout Manifest Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "review_kind": {
            "enum": [
              "overlap",
              "storage",
              "isolation",
              "manifest"
            ],
            "title": "Review Kind",
            "type": "string"
          },
          "reviewer_user_id": {
            "format": "uuid",
            "title": "Reviewer User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "holdout_manifest_id",
          "reviewer_user_id",
          "review_kind",
          "decision",
          "evidence_digest",
          "reason",
          "created_at"
        ],
        "title": "HoldoutReviewSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AI Engineer Bench API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/maintainer/holdouts": {
      "post": {
        "operationId": "create_holdout_v1_maintainer_holdouts_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HoldoutManifestCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldoutManifestSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Holdout",
        "tags": [
          "maintainer-holdouts"
        ]
      }
    },
    "/v1/maintainer/holdouts/{holdout_id}": {
      "get": {
        "operationId": "get_holdout_v1_maintainer_holdouts__holdout_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "holdout_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Holdout Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldoutManifestSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Holdout",
        "tags": [
          "maintainer-holdouts"
        ]
      }
    },
    "/v1/maintainer/holdouts/{holdout_id}/freeze": {
      "post": {
        "operationId": "freeze_holdout_v1_maintainer_holdouts__holdout_id__freeze_post",
        "parameters": [
          {
            "in": "path",
            "name": "holdout_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Holdout Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldoutManifestSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Freeze Holdout",
        "tags": [
          "maintainer-holdouts"
        ]
      }
    },
    "/v1/maintainer/holdouts/{holdout_id}/retire": {
      "post": {
        "operationId": "retire_holdout_v1_maintainer_holdouts__holdout_id__retire_post",
        "parameters": [
          {
            "in": "path",
            "name": "holdout_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Holdout Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldoutManifestSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retire Holdout",
        "tags": [
          "maintainer-holdouts"
        ]
      }
    },
    "/v1/maintainer/holdouts/{holdout_id}/reviews": {
      "post": {
        "operationId": "review_holdout_v1_maintainer_holdouts__holdout_id__reviews_post",
        "parameters": [
          {
            "in": "path",
            "name": "holdout_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Holdout Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HoldoutReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldoutReviewSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Holdout",
        "tags": [
          "maintainer-holdouts"
        ]
      }
    },
...
  }
}